moved lomap scorers into own file
